Add additional tests for mock data generation functions
